music: add new Requester struct to encapsulate and extend User
The User struct contains a bare amount of information about a user, probably
that which won't change too frequently. As queue/autoplay somewhat strongly
depend on caching everything in lists, collect the data at the time of
enqueue, register, info and slap it in Requester.

This can probably be rewritten in the future to be macroized to avoid needing
another async function in not-particularly async code.

Alternatively, there can be an option in the future to obtain this information
at runtime somehow instead, perhaps in the actually async code that will send
out the final message.