autoplay: add command to reset all autoplay usertime scores to zero
In the future, non-autoplay queued songs might be configurable to count
towards usertime, so having a way to reset to zero might be useful in the
event a user royally screws up their score.

Future revisions of this command should consider implementing different
methods on how to rebalance the score, perhaps via normalization, average,
etc.

For now, this might be useful in debug scenarios, or working around bugs.