overhaul: replace voice_ready check with in_same_voice and join_voice!
voice_ready was a convenient way in the beginning of the project to make sure
that the bot was in voice for certain commands, and if it wasn't, automatically
join.

Turns out, !help calls checks, and thus it makes life hard for everyone.
This overhaul splits voice_ready into two, and incidentally fixes the
stale voice bug.

in_same_voice is the check that replaces voice_ready in many scenarios, and
simply checks that the command user is in the same voice channel as the bot.
It does NOT join the voice channel.

join_voice! is a macro that should be called BEFORE get_mstate!, and will
attempt to join the voice channel that the command caller is in. If the
command caller is not in voice, or in a different channel, it handles
replying with an error message and terminates the command early.

The stale voice bug was accidentally fixed by adding an additional check
in join_voice! on if mstate had been initialized (in voice) yet or not.
Restarting the bot quickly will keep the voice state, but mstate will
not be initialized, hence why nothing could play.