refactor part 4: remove final bits of discord-related types from MusicState
So, the objective of this commit was to move the "sticky message" from MusicState to
the discord-specific DiscordPlayer. This kind of revealed a larger problem with
MusicState containing the MusicPlayer -- will need to access the DiscordPlayer struct
itself at some point to be able to get the embedded data.

This commit steps around this problem by templating MusicState rather than having it
reference the DiscordPlayer by the MusicPlayer trait.

In the future (and perhaps a later refactor commit in this series), perhaps the
ownership of the two states should be flipped: a MusicPlayer should have a reference
to a MusicState struct, so music controls can be abstracted in the MusicPlayer trait,
and MusicStates do not need to carry spooky references to generics/dyn traits/whatever.