refactor part 5.2: further rework requester and user manager in autoplay
In the attempts to decouple MusicState/Autoplay, make Requester hold an enum
for the backend-specific User types. Templates could have been used here, but
in the event that multiple frontends might be able to control a single music
backend, I'd rather not template everything now and then un-template everything
again.

Not a huge fan of the enum approach though, might have to change that or use
some kind of better trait bound approach that avoids needing a centralized
list of all user model types.