refactor part 6.5?: fix webdash builds by moving webdata Intos into music
Turns out these Into impls are continuing to be a problem. webdata depending on
music pulls in a ton of extra garbage due to music's deps. It would be better
if webdata depended on literally no other deps, and the communication bits
are handled by the endpoints.

Mildly confirms that either webdata definitely needs to stay, or needs to be
folded into music under a `models` feature, so importing music can be
relatively lightweight.