big refactor: remove the template off MusicState, replace with message passing
This is a chunky commit, strap in.

MusicState being a template made a mess of everywhere it had to be read from.
Since the object is to abstract away the frontends and music playing backends,
there should be no need to have knowledge of a specific MusicPlayer backend
anymore, and having it be a dyn Trait reference instead is even more of a pain.

Instead, MusicState now contains an (at the moment) mpsc sender, which it issues
commands via message passing to the music player, and the music player will do
its own thing, and send back via oneshot the Result<>. This is effectively the
same thing, but now airgaps the two components such that MusicState no longer needs
a direct reference to the player, and thus neither does any downstream code that
holds a reference to MusicState.

As a side of effect to this overhaul, the web frontend has been disabled temporarily.
It was some of the only code that touched the music player directly, and since it
will be factored out of the discord crate eventually, there is no purpose in
updating it now.

A lot of the features in main.rs were also tossed out. The initialization flow
needs an overhaul as well, especially with the new message passing method of
handling component interaction.