web overhaul: move the webapi to its own crate
The webapi should be its own crate, independent of discord. In order to
make this relatively breaking change, MusicState is now in charge of the
broadcasts, rather than the discord client code.

This is closer to what it should be in the end, ideally any mutates to
MusicState should probably broadcast to frontends that something changed.
Ideally, both the discord and webdash frontends should handle these
broadcasts, not just web.