music: create a songended handler on MusicState to be called by a player
In the effort to make new players/frontends easier to spin up, the "next song"
behavior has been moved into MusicState. For now, it is a function that can be
called on mstate when a song ends.

This will very likely be changed in the future, where instead MusicState will
listen on an mpsc for SongEnd events, and update itself that way. That way,
players (and maybe frontends too?) will be able to signal more events to
mstate as a whole, perhaps removing the need for a reference to mstate in
the first place.