main: replace the infinite loop with awaiting MusicState as suggested by clippy
Apparently, loop{} wastes cpu cycles. So for now, await on MusicState's
event loop, which should never terminate (at the moment). This should be
replaced by a top-level application manager, that looks for signals to
start/stop certain services, or the whole thing.

Note for the future: check if this improves music jitter, or if this
introduces more crashes due to unhandled unwraps.