webdash/progress: mitigate bug where progress bar reset on broadcast
The nowplaying progress bar recreates its Interval every time the
component changes. This probably helps with drift a bit, but the
main purpose is to start it again if it is canceled from reaching
the end.

However, when the initial progress support (intended for late-join
or page refresh) was added, it now causes a component refresh for
every broadcast. This fix just sets the recreated interval's time
to the new broadcasted progress (and probably adjusts drift a bit).

Better solution ideas moving forward:
 - Partial broadcasts will avoid rerendering nowplaying since it will
  just update the relevant components
 - Don't recreate the interval, just always let it tick. Consider using
  a faster tick rate to avoid noticeable offset
 - Use some on-first-render logic, or partial prop change checks to
  avoid a rerender if only progress changed