overhaul: replace music::Song with model::Song, add composite SongRequest
In the neverending quest to make things more database friendly, music::Song
and model::Song have been unified into the latter. model::Song is now
specifically just information about the Song. The requested_by field has
been moved up into a containing struct called SongRequest, which is just
a Song and a Requester.

This is part of the effort to make things more serializable -- ideally Song
and Requester information do not need to be directly tied to each other,
that is, in the same struct. They will likely come from different locations,
that is, different tables in a database.

Therefore, it would be easier to serialize directly into invidiual structs
if they appear similar. Furthermore, it allows for the creation of Song
structs without needing a Requester, which has been a slight pain point
in some areas, when it would be useful for just simple rendering.