massive storage overhaul part 1, replace pickledb with an actual db
Persistent storage has now been replaced with an actual sqlite database.
This is a massive work in progress over the next several commits. As of
this commit, it is read-only -- data must be manually inserted into the
db. To keep scope relatively small, this ONLY replicates the original
behavior of the original pickledb approach, albeit with some smaller
included refactors to avoid writing throwaway code for compatibility.

Some changes include:
 - use MinstrelUserId everywhere. This is now the primary key of the
 user column, instead of being the UserId from discord.
 - playlist urls are now replaced with a Source enum. Ideally, this
 should make adding future source types easier maybe.
 - a lot of discord helpers are now dependent on the db, expect some
 refactors to passing application state everywhere