storage part 2: add a new UserMgmt "adapter", for higher-level user functions
This probably should have been split up, but in order to make sure it mostly
worked correctly, it made sense to write an implementation first.

UserMgmt is a new "adapter" on the to-be-renamed MusicAdapter that provides
high level user management controls, such as creating a new user, and
methods to link one user auth method to another (i.e. user created with
discord to link to a web auth)

User merging is something to figure out later, a TODO stub has been left
in the code.

For now, user metadata is used from first registration, expect future
API functions to allow modification.