music: change .get_adapter() to return a clone of a central adapter
Adapters are all meant to share the same state with each other, however
calling the ::new() constructor on it when generating each new adapter
means that some constructors of sub-adapters are going to be called
again, and probably generate new internal state.

This might be solved in the future when ideally the adapters are all
somewhat unified, but for now just try to avoid putting too much
shared state in a single struct again, that was the point of avoiding
a large Arc/Mutex'd state struct in the first place.