music/musicstate: fixed an esoteric error where the music state was set to stopped despite a song being played
For some reason, .song_ended() was setting state = Stopped if .next() was *OK*.

This is entirely wrong.

.next() sets the state to Playing via a call to .play(), then returns whatever .play() returns. If there is
nothing to play though, it returns EmptyQueue. I assume this is what that `if` was *supposed* to check for.
Replace the whole thing with a match, and included an additional warning for other responses that are currently
not expected.