music: don't fire a SongEnded event when the Player reports back a playback failure
If a Player can't play a song due to some issue with the Player, it will likely
fail like that again on the next song. However, a previous hack fired an extra
SongEnd event so that if a song failed due to say, unavailability, it wouldn't
just completely stop the music playback entirely.

This is a holdover fix to at least stop the bot sprinting through every song
autoplay can come up with by giving up if the Player reports back a
PlaybackFailed error. Perhaps this should instead always fail, and only try
to continue if in 1. a mode where it should keep playing and 2. Player
reports an error specific to that song, and not indicative of a larger problem.