music: update youtube_dl dependency, filter out songs from playlists without durations
Either an update to yt-dlp or the youtube_dl crate now shows deleted
videos from playlists. In this case, most fields in the SingleVideo
struct are None.

Filter out any songs without a duration, since we will panic on that
value later anyway. This may be temporary in lieu of a better way later
to filter out deleted songs.