music/song: use channel instead of uploader to approximate the artist
Apparently the yt-dlp output may have changed, and now uploader seems
to be empty on acquisition of a playlist. Perhaps in the future there
should be an attempt to fetch metadata from an alternate source, or from
cache.

For now, use channel instead. Some artist names are likely going to
change from what they were displayed as before, some may still be
Unknown.